chore: move current to its own file
